@@ -1,5 +1,3 @@
-# -*- coding: utf-8 -*-
-#
 # Downward Lab uses the Lab package to conduct experiments with the
 # Fast Downward planning system.
 #
@@ -22,7 +20,7 @@ import os.path
 from lab import tools
 
 
-class ScatterPgfplots(object):
+class ScatterPgfplots:
     @classmethod
     def _get_plot(cls, report):
         lines = []
@@ -124,10 +122,10 @@ class ScatterPgfplots(object):
                 continue
             if isinstance(value, bool) or value is None:
                 opts.append(key)
-            elif isinstance(value, tools.string_type):
+            elif isinstance(value, str):
                 if " " in value or "=" in value:
                     value = "{%s}" % value
                 opts.append("{}={}".format(key, value.replace("_", "-")))
             else:
-                opts.append("{}={}".format(key, value))
+                opts.append(f"{key}={value}")
         return ", ".join(opts)
